[refactor] BaseService 구조 분리
